added user service methods for login and logout
